+++ tests/fixtures/cloudintegrations.py
