+++ tests/fixtures/cloudintegrationsutils.py
