+++ tests/fixtures/fingerprint.py
