+++ tests/fixtures/gatewayutils.py
